Fixed a crash caused by falling blocks

+++ src/main/java/mekanism/generators/common/tile/reactor/TileEntityReactorPort.java
@@ -218,7 +218,7 @@ public class TileEntityReactorPort extends TileEntityReactorBlock implements IFl
 
     @Override
     public double getMaxOutput() {
-        return Integer.MAX_VALUE;
+        return Long.MAX_VALUE;
     }
 
     @Override
